Develop

+++ paprika-web/src/main/java/app/controller/FormController.java
@@ -22,6 +22,75 @@ import app.utils.ViewUtil;
 public class FormController {
 
 	/**
+	 * Standard Page for reset password, configure for send.
+	 */
+	public static final Route serveFormResetSendPage = (Request request, Response response) -> {
+
+		Map<String, Object> model = new HashMap<>();
+		PaprikaWebMain.LOGGER.trace("-------serveFormResetPage--------");
+
+		model.put("send", true);
+		return ViewUtil.render(request, model, PathIn.Template.RESET);
+	};
+	/**
+	 * Standard Page for reset password, configure for change.
+	 */
+	public static final Route serveFormResetReceivePage = (Request request, Response response) -> {
+
+		Map<String, Object> model = new HashMap<>();
+		PaprikaWebMain.LOGGER.trace("-------serveFormResetPage--------");
+		model.put("change", true);
+
+		return ViewUtil.render(request, model, PathIn.Template.RESET);
+	};
+
+	/**
+	 * POST page for resetpassword. If do  not contains pwd on the request, so this is a post from of a SEND page, else CHANGE page.
+	 * 
+	 * Merged for obscur reason.
+	 */
+	public static final Route handleFormResetPost = (Request request, Response response) -> {
+		Map<String, Object> model = new HashMap<>();
+
+		PaprikaWebMain.LOGGER.trace("-------handleFormResetPost--------");
+		String email = RequestUtil.getQueryUsername(request);
+
+		String captcha = request.queryParams("g-recaptcha-response");
+		boolean stop = false;
+		if (PaprikaWebMain.ENABLEALLSECURITY)
+			if (captcha == null || captcha.isEmpty() || "false".equals(captcha)) {
+				model.put("resetFlagFail", true);
+				stop = true;
+			}
+		if (email != null) {
+			PaprikaFacade facade = PaprikaFacade.getInstance();
+			String pwd = RequestUtil.getQueryPassword(request);
+
+			if (pwd != null) {
+
+				model.put("change", true);
+				if (stop)
+					return ViewUtil.render(request, model, PathIn.Template.RESET);
+
+				String activation = request.queryParams("activation");
+				boolean flag = facade.resetpwd(email, activation, pwd);
+				model.put("resetFlag", flag);
+
+			} else {
+
+				model.put("send", true);
+				if (stop)
+					return ViewUtil.render(request, model, PathIn.Template.RESET);
+				facade.sendnewPwd(email);
+				model.put("emailSended", true);
+			}
+		}
+		model.put(PaprikaKeyWords.PROJECT, null);
+
+		return ViewUtil.render(request, model, PathIn.Template.RESET);
+	};
+
+	/**
 	 * If a user try to go on the page without be logged. Go on Login page.
 	 * Else, nothing.
 	 */
@@ -32,7 +101,7 @@ public class FormController {
 
 		LoginController.ensureUserIsLoggedIn(request, response);
 		Map<String, Object> model = new HashMap<>();
-		PaprikaWebMain.LOGGER.trace("-------handleFormDelPage--------");
+		PaprikaWebMain.LOGGER.trace("-------serveFormDELPage--------");
 		return ViewUtil.render(request, model, PathIn.Template.FORM_DELETE);
 	};
 
@@ -43,17 +112,24 @@ public class FormController {
 	public static final Route handleFormDeletePost = (Request request, Response response) -> {
 		Map<String, Object> model = new HashMap<>();
 
-		//Project project = RequestUtil.getSessionProject(request);
+		// Project project = RequestUtil.getSessionProject(request);
 		PaprikaWebMain.LOGGER.trace("-------handleFormDeletePost--------");
 		String delete = request.queryParams("delete");
-		if (delete != null) {
+		
+		String captcha = request.queryParams("g-recaptcha-response");
+		if (PaprikaWebMain.ENABLEALLSECURITY)
+			if (captcha == null || captcha.isEmpty() || "false".equals(captcha)) {
+				model.put("resetFlagFail", true);
+				return ViewUtil.render(request, model, PathIn.Template.FORM_DELETE);
+			}
+		
+		if (delete != null  && !delete.isEmpty()) {
 			PaprikaWebMain.LOGGER.trace("etape delete: " + delete);
-			deleteNotNull(request, null);//project);
-			//model.put(PaprikaKeyWords.PROJECT, project);
+			deleteNotNull(request, null);// project);
+			// model.put(PaprikaKeyWords.PROJECT, project);
 		}
 		model.put(PaprikaKeyWords.PROJECT, null);
-		
-		
+
 		return ViewUtil.render(request, model, PathIn.Template.FORM_DELETE);
 	};
 
@@ -66,14 +142,17 @@ public class FormController {
 		Set<String> setOfIdToDelete = new HashSet<>();
 		for (String params : setQueryParams) {
 			String idtoDelete = request.queryParams(params);
-			if (idtoDelete == null || "00-".equals(idtoDelete)) {
+			if (idtoDelete == null || "00-".equals(idtoDelete) || idtoDelete.isEmpty() ) {
 				continue;
 			}
-			setOfIdToDelete.add(idtoDelete);
+			if(idtoDelete.startsWith("00-")){
+				
+			setOfIdToDelete.add(idtoDelete.substring(3));
+			}
 		}
 		PaprikaFacade facade = PaprikaFacade.getInstance();
 		facade.deleteOnDataBase(setOfIdToDelete);
-		//facade.needReloadApp(project);
+		// facade.needReloadApp(project);
 	}
 
 }
